Add the vue template compiler onto any template loader specified in the vue.loaders configuration (#595)
* Tweak to add the vue template compiler onto any loader specified in the vue.loaders configuration.

This enables, for example, using a markdown loader via <template lang="md"> in a .vue file.  Without this tweak, the webpack config has
to specify both the loader that converts the markdown (e.g.) to HTML AND template-compiler to compile it.  That, however doesn't fully work
because the webpack config can't know the moduleId generated by lib/loader.js and pass it in via the `id` parameter, so scoped styles won't
work.

* adjustments

* add tests for custom html loader